backend/requirements.txt
fastapi
uvicorn
pyjwt
redis
fastapi-limiter
SQLAlchemy
psycopg2-binary
alembic
